Move logic to story editor package.
