+++ packages/element-library/package.json
@@ -53,7 +53,6 @@
     "classnames": "^2.3.1",
     "polished": "^4.2.2",
     "prop-types": "^15.7.2",
-    "react-blurhash": "^0.1.3",
     "react-transition-group": "^4.4.2",
     "styled-components": "^5.2.1"
   },
